refactor: 💡 PR Ajoute le menu ainsi que les CGU et ML
✅ Closes: #226
