+++ front/components/form/DatePicker.tsx
@@ -21,7 +21,7 @@ import {
   PLATFORM_IS_IOS,
 } from "../../constants/platform.constants";
 import Shadow from "../../constants/shadow.constants";
-import Icomoon, { IcomoonIcons } from "../Icomoon";
+import Icomoon, { IcomoonIcons } from "../icomoon.component";
 import Button from "./Button";
 
 interface Props {
