+++ front/components/index.ts
@@ -12,14 +12,18 @@ import Button from "../components/form/Button";
 import Checkbox from "../components/form/Checkbox";
 import Datepicker from "../components/form/DatePicker";
 import HeaderApp from "../components/HeaderApp";
-import Icomoon from "../components/Icomoon";
+import Modal from "../components/modal.component";
 import { CommonText, SecondaryText } from "../components/StyledText";
 import StepIcon from "../components/timeline/StepIcon";
 import TimlineStep from "../components/timeline/TimlineStep";
+import Backdrop from "./backdrop.component";
+import Icomoon, { IcomoonIcons } from "./icomoon.component";
+import Menu from "./menu.component";
 
 export {
   Agenda,
   BackButton,
+  Backdrop,
   Button,
   Checkbox,
   CommonText,
@@ -27,9 +31,12 @@ export {
   DidYouKnow,
   HeaderApp,
   Icomoon,
+  IcomoonIcons,
   ImageBanner,
   InShort,
   Links,
+  Menu,
+  Modal,
   SecondaryText,
   StepIcon,
   SubTitle,
