+++ front/screens/ConditionsOfUse.tsx
@@ -2,7 +2,7 @@ import type { FC } from "react";
 import * as React from "react";
 
 import htmlFile from "../assets/html/cgu.html";
-import Modal from "../components/Modal";
+import { Modal } from "../components";
 
 interface Props {
   setIsVisible: (showMenu: boolean) => void;
